+++ tests/Plugin/Wechat/GeneralPluginTest.php
@@ -12,10 +12,7 @@ use Yansongda\Pay\Tests\TestCase;
 
 class GeneralPluginTest extends TestCase
 {
-    /**
-     * @var \Yansongda\Pay\Tests\Stubs\Plugin\WechatGeneralPluginStub
-     */
-    protected $plugin;
+    protected WechatGeneralPluginStub $plugin;
 
     protected function setUp(): void
     {
@@ -36,6 +33,7 @@ class GeneralPluginTest extends TestCase
         self::assertInstanceOf(RequestInterface::class, $radar);
         self::assertEquals('POST', $radar->getMethod());
         self::assertEquals(new Uri(Wechat::URL[Pay::MODE_NORMAL].'yansongda/pay'), $radar->getUri());
+        self::assertEquals('mp_app_id', $result->getPayload()['config_key']);
     }
 
     public function testPartner()
@@ -50,5 +48,20 @@ class GeneralPluginTest extends TestCase
         self::assertInstanceOf(RequestInterface::class, $radar);
         self::assertEquals('POST', $radar->getMethod());
         self::assertEquals(new Uri(Wechat::URL[Pay::MODE_SERVICE].'yansongda/pay/partner'), $radar->getUri());
+    }
+
+    public function testGetConfigKey()
+    {
+        $rocket = new Rocket();
+        $rocket->setParams(['_type' => 'mini']);
+
+        $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
+
+        $radar = $result->getRadar();
+
+        self::assertInstanceOf(RequestInterface::class, $radar);
+        self::assertEquals('POST', $radar->getMethod());
+        self::assertEquals(new Uri(Wechat::URL[Pay::MODE_NORMAL].'yansongda/pay'), $radar->getUri());
+        self::assertEquals('mini_app_id', $result->getPayload()['config_key']);
     }
 }