+++ tests/Plugin/Wechat/RadarSignPluginTest.php
@@ -7,24 +7,24 @@ use ReflectionClass;
 use Yansongda\Pay\Contract\ConfigInterface;
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Exception\InvalidConfigException;
+use Yansongda\Pay\Packer\JsonPacker;
+use Yansongda\Pay\Packer\XmlPacker;
 use Yansongda\Pay\Pay;
 use Yansongda\Pay\Plugin\Wechat\RadarSignPlugin;
 use Yansongda\Pay\Rocket;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
+use Yansongda\Supports\Str;
 
 class RadarSignPluginTest extends TestCase
 {
-    /**
-     * @var \Yansongda\Pay\Plugin\Wechat\RadarSignPlugin
-     */
-    protected $plugin;
+    protected RadarSignPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new RadarSignPlugin();
+        $this->plugin = new RadarSignPlugin(new JsonPacker(), new XmlPacker());
     }
 
     public function testNormal()
@@ -42,7 +42,7 @@ class RadarSignPluginTest extends TestCase
 
         self::assertTrue($radar->hasHeader('Authorization'));
         self::assertFalse($radar->hasHeader('Wechatpay-Serial'));
-        self::assertEquals(json_encode($params), $radar->getBody()->getContents());
+        self::assertEquals(json_encode($params), (string) $radar->getBody());
     }
 
     public function testNormalWithWechatSerial()
@@ -82,7 +82,7 @@ class RadarSignPluginTest extends TestCase
         $contents = "POST\n/v3/pay/transactions/h5\n1626493236\nQqtzdVzxavZeXag9G5mtfzbfzFMf89p6\n{\"out_trade_no\":1626493236,\"description\":\"yansongda 测试 - 1626493236\",\"amount\":{\"total\":1},\"scene_info\":{\"payer_client_ip\":\"127.0.0.1\",\"h5_info\":{\"type\":\"Wap\"}},\"appid\":\"wx55955316af4ef13\",\"mchid\":\"1600314069\",\"notify_url\":\"http:\/\/127.0.0.1:8000\/wechat\/notify\"}\n";
 
         $class = new ReflectionClass($this->plugin);
-        $method = $class->getMethod('getWechatAuthorization');
+        $method = $class->getMethod('v3GetWechatAuthorization');
         $method->setAccessible(true);
 
         $result = $method->invokeArgs($this->plugin, [$params, $timestamp, $random, $contents]);
@@ -119,7 +119,7 @@ class RadarSignPluginTest extends TestCase
         self::expectExceptionMessage('Missing Wechat Config -- [mch_public_cert_path]');
 
         $class = new ReflectionClass($this->plugin);
-        $method = $class->getMethod('getWechatAuthorization');
+        $method = $class->getMethod('v3GetWechatAuthorization');
         $method->setAccessible(true);
         $method->invokeArgs($this->plugin, [$params, $timestamp, $random, $contents]);
     }
@@ -150,8 +150,30 @@ class RadarSignPluginTest extends TestCase
         self::expectExceptionMessage('Parse [mch_public_cert_path] Serial Number Error');
 
         $class = new ReflectionClass($this->plugin);
-        $method = $class->getMethod('getWechatAuthorization');
+        $method = $class->getMethod('v3GetWechatAuthorization');
         $method->setAccessible(true);
         $method->invokeArgs($this->plugin, [$params, $timestamp, $random, $contents]);
+    }
+
+    public function testV2Normal()
+    {
+        $params = [
+            'name' => 'yansongda',
+            'age' => 28,
+        ];
+        $rocket = (new Rocket())->setParams(array_merge($params, ['_version' => 'v2']))
+                                ->setPayload(new Collection($params))
+                                ->setRadar(new Request('POST', '127.0.0.1'));
+
+        $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
+
+        $radar = $result->getRadar();
+        $payload = $result->getPayload();
+
+        self::assertTrue($payload->has('sign'));
+        self::assertTrue($payload->has('nonce_str'));
+        self::assertTrue(Str::contains((string) $radar->getBody(), '<xml>'));
+        self::assertTrue(Str::contains((string) $radar->getBody(), '<sign>'));
+        self::assertTrue(Str::contains((string) $radar->getBody(), '<nonce_str>'));
     }
 }