+++ tests/Stubs/Plugin/WechatGeneralPluginStub.php
@@ -9,6 +9,7 @@ class WechatGeneralPluginStub extends GeneralPlugin
 {
     protected function doSomething(Rocket $rocket): void
     {
+        $rocket->mergePayload(['config_key' => $this->getConfigKey($rocket->getParams())]);
     }
 
     protected function getUri(Rocket $rocket): string