+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## TBD-v3.3.0
 
+### added
+
+- feat: 支持微信 v2 版本刷卡支付(#753)
+
 ### deleted
 
 - delete: 移除废弃的类(#752)
@@ -12,7 +16,8 @@
 
 - change: 所有的 `Find*Plugin` 调整为 `Query*Plugin`(#756)
 - change: 插件开始装载日志由 `info` 调整为 `debug`(#755)
-- change: ParserInterface 签名由 `?ResponseInterface $response` 调整为 `PackerInterface $packer, ?ResponseInterface $response`(#754)
+- change: ParserInterface 签名由 `?ResponseInterface $response` 变更为 `PackerInterface $packer, ?ResponseInterface $response`(#754)
+- change: \Yansongda\Pay\Plugin\Wechat\RadarSignPlugin 增加 `__construct(JsonPacker $jsonPacker, XmlPacker $xmlPacker)` 方法(#753)
 
 ## v3.2.14